backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.1
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
marshmallow==3.21.3
Pillow==10.3.0
gunicorn==22.0.0
eventlet==0.36.1
pytest==8.2.2
pytest-flask==1.3.0
flasgger==0.9.7.1
